requirements.txt
fastapi==0.95.0
uvicorn==0.18.2
psycopg2-binary==2.9.3
pandas==1.5.3
numpy==1.23.3